Modify run_mulm-perm.py

+++ ml/mulm_gpu/mulm/run_mulm-perm.py
@@ -1,6 +1,5 @@
 #! /usr/bin/env python
 
-#from __future__ import division
 import swig.linear_model as lm
 import os.path
 import iodata
@@ -34,7 +33,7 @@ TRESHOLD = 10
 cut = 90
 divide = 4 * (int(num_permut*((100.0-cut)/100.0))+1)
 print divide
-Beta_Size=long(num_vector_X*num_vector_Y*divide)
+Beta_Size=num_vector_X*num_vector_Y*divide
 print Beta_Size
 X=(X.T).reshape(num_vector_X*VectorSize)
 X=X.astype('float32')
@@ -52,7 +51,6 @@ for i in range(num_permut-1):
 	np.random.shuffle(indx)
 	P=np.append(P,indx)
 P=P.astype('int32')
-#Beta = lm.create_array(Beta_Size)
 values , Beta = lm.MULMRegression(X, Y, Z, P,Beta_Size, VectorSize, divide, TRESHOLD, dev)
 
 print values
